hfcs-fffit-main/requirements-conda.txt
signac
signac-flow
unyt
mbuild<0.11.0
foyer
git-lfs
mosdef_cassandra<=0.3.0
panedr
matplotlib
pandas
seaborn
scikit-learn
numpy>=1.19.2,<1.20.0
